+++ packages/frontend/src/components/VisualizationConfigs/PieChartConfig/PieChartConfigTabs.tsx
@@ -1,9 +1,9 @@
 import { MantineProvider, Tabs } from '@mantine/core';
 import React, { memo } from 'react';
-import { themeOverride } from '../../mantineTheme';
-import PieChartDisplayConfig from '../PieChartDisplayConfig';
-import PieChartLayoutConfig from '../PieChartLayoutConfig';
-import PieChartSeriesConfig from '../PieChartSeriesConfig';
+import { themeOverride } from '../mantineTheme';
+import PieChartDisplayConfig from './PieChartDisplayConfig';
+import PieChartLayoutConfig from './PieChartLayoutConfig';
+import PieChartSeriesConfig from './PieChartSeriesConfig';
 
 export const ConfigTabs: React.FC = memo(() => {
     return (
